+++ crates/base/macros/src/yard/mod.rs
@@ -1,3 +1,3 @@
-// devela_baes_macros::yard
+// devela_base_macros::yard
 
 // pub mod _dep;