+++ crates/base/std/src/num/dom/real/float/wrapper/mod.rs
@@ -5,10 +5,10 @@
 
 mod definition; // FloatStd
 
-mod basic; // basic operations SYMLINK to /src/base/std/src/num/dom/real/float/wrapper/basic.rs
-mod consts; // constants       SYMLINK to /src/base/std/src/num/dom/real/float/wrapper/consts.rs
-mod minimax; // Horner-rel fns SYMLINK to /src/base/std/src/num/dom/real/float/wrapper/minimax.rs
-mod series; // Taylor-rel fns  SYMLINK to /src/base/std/src/num/dom/real/float/wrapper/series.rs
+mod basic; // basic operations SYMLINK to /crates/base/std/src/num/dom/real/float/wrapper/basic.rs
+mod consts; // constants       SYMLINK to /crates/base/std/src/num/dom/real/float/wrapper/consts.rs
+mod minimax; // Horner-rel fns SYMLINK to /crates/base/std/src/num/dom/real/float/wrapper/minimax.rs
+mod series; // Taylor-rel fns  SYMLINK to /crates/base/std/src/num/dom/real/float/wrapper/series.rs
 
 mod std; // std methods