+++ src/_doc/macros/mod.rs
@@ -1,0 +1,6 @@
+// devela::_doc::macros
+//
+// #![doc = crate::_tags!(procedural_macro)]
+//! Procedural macros.
+//!
+#![doc = include_str!("./devela_base_macros.md")]